+++ packages/amplify-graphql-api-construct-tests/src/__tests__/sql-pg-models.test.ts
@@ -1,4 +1,4 @@
-import { createNewProjectDir, deleteProjectDir } from 'amplify-category-api-e2e-core';
+import { createNewProjectDir, deleteProjectDir, generateDBName } from 'amplify-category-api-e2e-core';
 import generator from 'generate-password';
 import { getResourceNamesForStrategyName } from '@aws-amplify/graphql-transformer-core';
 import { SqlDatatabaseController } from '../sql-datatabase-controller';
@@ -12,11 +12,10 @@ describe('CDK GraphQL Transformer deployments with Postgres SQL datasources', ()
   let projRoot: string;
   const projFolderName = 'pgmodels';
 
-  const [username, password, identifier] = generator.generateMultiple(3);
-
+  // sufficient password length that meets the requirements for RDS cluster/instance
+  const [username, password, identifier] = generator.generateMultiple(3, { length: 11 });
   const region = process.env.CLI_REGION ?? 'us-west-2';
-
-  const dbname = 'default_db';
+  const dbname = generateDBName();
   const engine = 'postgres';
 
   const databaseController: SqlDatatabaseController = new SqlDatatabaseController(