+++ packages/amplify-data-construct/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.9.3...@aws-amplify/data-construct@1.9.4) (2024-08-01)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
+## [1.9.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.9.2...@aws-amplify/data-construct@1.9.3) (2024-07-25)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
 ## [1.9.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.9.1...@aws-amplify/data-construct@1.9.2) (2024-07-15)
 
 **Note:** Version bump only for package @aws-amplify/data-construct